fix: Rename command validator, add validation err
